Optional automatic daily log file rotation
This is probably most useful on windows (cygwin) where
a convenient daily log rotation tool may not be available.

Every day the log file is re-opened and the log file name
pre-parsed with strftime(), so using items such as %d
in the file name will then create a new file each day. See the
strftime() help page for full list of possibilities,
%Y-%m-%d for example could be used to add an ISO8601 date
string to the daily log file name.